miku/requirements.txt
discord.py==2.1.1
requests
# python-dotenv
motor

